+++ model/message.go
@@ -1,7 +1,7 @@
 package model
 
 type Message struct {
-	ID       uint64 `genji:"pk"`
+	ID       uint32 `genji:"pk"`
 	DateUTC  int64  `genji:"index"`
 	FlagsCSV string
 }
