+++ model/message_body_raw.go
@@ -1,6 +1,6 @@
 package model
 
 type MessageBodyRaw struct {
-	ID   uint64 `genji:"pk"`
+	ID   uint32 `genji:"pk"`
 	Body []byte
 }
